refactor - now using preprocessor directives to control net40 differences
consolidated projects into a single solution
refactor of sink options into base class
