+++ serilog-sinks-amazonkinesis.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{037440DE-440B-4129-9F7A-09B42D00397E}"
 EndProject
@@ -23,6 +23,10 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis", "src\Serilog.Sinks.Amazon.Kinesis\Serilog.Sinks.Amazon.Kinesis.csproj", "{9CDDC147-93BB-47DC-899C-B41384D7AE23}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis.Firehose", "src\Serilog.Sinks.Amazon.Kinesis.Firehose\Serilog.Sinks.Amazon.Kinesis.Firehose.csproj", "{81B149AE-B673-401B-9E7E-F7B56F6B8BED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis.Firehose-net40", "src\Serilog.Sinks.Amazon.Kinesis.Firehose\Serilog.Sinks.Amazon.Kinesis.Firehose-net40.csproj", "{00297150-F24F-43BF-A681-1A3FC5551880}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Amazon.Kinesis.Stream-net40", "src\Serilog.Sinks.Amazon.Kinesis.Stream\Serilog.Sinks.Amazon.Kinesis.Stream-net40.csproj", "{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,6 +50,14 @@ Global
 		{81B149AE-B673-401B-9E7E-F7B56F6B8BED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81B149AE-B673-401B-9E7E-F7B56F6B8BED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{81B149AE-B673-401B-9E7E-F7B56F6B8BED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00297150-F24F-43BF-A681-1A3FC5551880}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00297150-F24F-43BF-A681-1A3FC5551880}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00297150-F24F-43BF-A681-1A3FC5551880}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00297150-F24F-43BF-A681-1A3FC5551880}.Release|Any CPU.Build.0 = Release|Any CPU
+		{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -55,5 +67,7 @@ Global
 		{6672FF6B-5243-4588-986A-EEA9C0BD6C0B} = {543E1315-71DE-40CE-A179-4B27B7A64125}
 		{9CDDC147-93BB-47DC-899C-B41384D7AE23} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{81B149AE-B673-401B-9E7E-F7B56F6B8BED} = {037440DE-440B-4129-9F7A-09B42D00397E}
+		{00297150-F24F-43BF-A681-1A3FC5551880} = {037440DE-440B-4129-9F7A-09B42D00397E}
+		{253E61AB-0DCF-4C6D-AC1E-DC6EB5529B10} = {037440DE-440B-4129-9F7A-09B42D00397E}
 	EndGlobalSection
 EndGlobal
