+++ src/Serilog.Sinks.Amazon.Kinesis.Firehose/Serilog.Sinks.Amazon.Kinesis.Firehose-net40.csproj
@@ -4,7 +4,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{81B149AE-B673-401B-9E7E-F7B56F6B8BED}</ProjectGuid>
+    <ProjectGuid>{00297150-F24F-43BF-A681-1A3FC5551880}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Serilog.Sinks.Amazon.Kinesis.Firehose</RootNamespace>
@@ -18,7 +18,7 @@
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
     <OutputPath>bin40\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <DefineConstants>DEBUG;TRACE;NET40</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
@@ -28,7 +28,7 @@
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>bin40\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
+    <DefineConstants>TRACE;NET40</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
@@ -63,11 +63,11 @@
     <Compile Include="App_Packages\LibLog.4.2\LibLog.cs" />
     <Compile Include="KinesisFirehoseLoggerConfigurationExtensions.cs" />
     <Compile Include="Sinks\DurableKinesisFirehoseSink.cs" />
-    <Compile Include="Sinks\HttpLogShipper-net40.cs" />
+    <Compile Include="Sinks\HttpLogShipper.cs" />
     <Compile Include="Sinks\KinesisFirehoseSinkOptions.cs" />
-    <Compile Include="Sinks\KinesisFirehoseSinkState.cs" />
     <Compile Include="Sinks\KinesisFirehoseSink.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Sinks\KinesisSinkState.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config">
