+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/KinesisLoggerConfigurationExtensions.cs
@@ -36,7 +36,7 @@ namespace Serilog
         /// <exception cref="ArgumentNullException">A required parameter is null.</exception>
         public static LoggerConfiguration AmazonKinesis(
             this LoggerSinkConfiguration loggerConfiguration,
-            KinesisSinkOptions options)
+            KinesisStreamSinkOptions options)
         {
             if (loggerConfiguration == null) throw new ArgumentNullException("loggerConfiguration");
             if (options == null) throw new ArgumentNullException("options");
@@ -85,12 +85,12 @@ namespace Serilog
             if (kinesisClient == null) throw new ArgumentNullException("kinesisClient");
             if (streamName == null) throw new ArgumentNullException("streamName");
 
-            var options = new KinesisSinkOptions(kinesisClient: kinesisClient, streamName: streamName, shardCount: shardCount)
+            var options = new KinesisStreamSinkOptions(kinesisClient: kinesisClient, streamName: streamName, shardCount: shardCount)
             {
                 BufferFileSizeLimitBytes = bufferFileSizeLimitBytes,
                 BufferBaseFilename = bufferBaseFilename,
-                Period = period ?? KinesisSinkOptions.DefaultPeriod,
-                BatchPostingLimit = batchPostingLimit ?? KinesisSinkOptions.DefaultBatchPostingLimit,
+                Period = period ?? KinesisStreamSinkOptions.DefaultPeriod,
+                BatchPostingLimit = batchPostingLimit ?? KinesisStreamSinkOptions.DefaultBatchPostingLimit,
                 MinimumLogEventLevel = minimumLogEventLevel ?? LevelAlias.Minimum,
                 OnLogSendError = onLogSendError
             };
