+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Sinks/KinesisApi.cs
