+++ src/Serilog.Sinks.Amazon.Kinesis.Stream/Sinks/KinesisSinkState.cs
@@ -20,24 +20,24 @@ namespace Serilog.Sinks.Amazon.Kinesis.Stream
 {
     internal class KinesisSinkState
     {
-        public static KinesisSinkState Create(KinesisSinkOptions options)
+        public static KinesisSinkState Create(KinesisStreamSinkOptions options)
         {
             if (options == null) throw new ArgumentNullException("options");
             var state = new KinesisSinkState(options);
             return state;
         }
 
-        private readonly KinesisSinkOptions _options;
+        private readonly KinesisStreamSinkOptions _options;
         private readonly IAmazonKinesis _client;
         private readonly ITextFormatter _formatter;
         private readonly ITextFormatter _durableFormatter;
 
-        public KinesisSinkOptions Options { get { return _options; } }
+        public KinesisStreamSinkOptions Options { get { return _options; } }
         public IAmazonKinesis KinesisClient { get { return _client; } }
         public ITextFormatter Formatter { get { return _formatter; } }
         public ITextFormatter DurableFormatter { get { return _durableFormatter; } }
 
-        private KinesisSinkState(KinesisSinkOptions options)
+        private KinesisSinkState(KinesisStreamSinkOptions options)
         {
             if (string.IsNullOrWhiteSpace(options.StreamName)) throw new ArgumentException("options.StreamName");
